refactor: move settings to general folder pkg
