+++ go.mod
@@ -1,3 +1,8 @@
 module github.com/GenesisEducationKyiv/software-engineering-school-4-0-Hukyl
 
 go 1.22.4
+
+require (
+	github.com/joho/godotenv v1.5.1
+	github.com/robfig/cron/v3 v3.0.1
+)
